New: Copying resource class parameters is opt-in
Previously, all parameters of a resource class were copied to the
resource value on initialization. This commit changes that so that
only the parameters explicitly specified are copied. This makes
more sense because some parameters won't be needed on the value
and keeping those in sync can take a lot of effort.